refactor(web/server): Migrate to Hypersistence Utils
This change updates the web server to use the Hypersistence Utils
library instead of our custom code to store columns as JSON in H2 and
Postgres.
