+++ opendc-web/opendc-web-server/src/main/kotlin/org/opendc/web/server/model/Portfolio.kt
@@ -23,9 +23,7 @@
 package org.opendc.web.server.model
 
 import org.hibernate.annotations.Type
-import org.hibernate.annotations.TypeDef
 import org.opendc.web.proto.Targets
-import org.opendc.web.server.util.hibernate.json.JsonType
 import javax.persistence.CascadeType
 import javax.persistence.Column
 import javax.persistence.Entity
@@ -45,7 +43,6 @@ import javax.persistence.UniqueConstraint
 /**
  * A portfolio is the composition of multiple scenarios.
  */
-@TypeDef(name = "json", typeClass = JsonType::class)
 @Entity
 @Table(
     name = "portfolios",
@@ -85,7 +82,7 @@ class Portfolio(
     /**
      * The portfolio targets (metrics, repetitions).
      */
-    @Type(type = "json")
+    @Type(type = "io.hypersistence.utils.hibernate.type.json.JsonType")
     @Column(columnDefinition = "jsonb", nullable = false, updatable = false)
     val targets: Targets
 ) {
