+++ opendc-web/opendc-web-server/src/main/kotlin/org/opendc/web/server/model/Topology.kt
@@ -23,9 +23,7 @@
 package org.opendc.web.server.model
 
 import org.hibernate.annotations.Type
-import org.hibernate.annotations.TypeDef
 import org.opendc.web.proto.Room
-import org.opendc.web.server.util.hibernate.json.JsonType
 import java.time.Instant
 import javax.persistence.Column
 import javax.persistence.Entity
@@ -43,7 +41,6 @@ import javax.persistence.UniqueConstraint
 /**
  * A datacenter design in OpenDC.
  */
-@TypeDef(name = "json", typeClass = JsonType::class)
 @Entity
 @Table(
     name = "topologies",
@@ -86,7 +83,7 @@ class Topology(
     /**
      * Datacenter design in JSON
      */
-    @Type(type = "json")
+    @Type(type = "io.hypersistence.utils.hibernate.type.json.JsonType")
     @Column(columnDefinition = "jsonb", nullable = false)
     var rooms: List<Room> = emptyList()
 ) {
