+++ opendc-web/opendc-web-server/src/main/resources/hypersistence-utils.properties
@@ -1,0 +1,1 @@
+hypersistence.utils.jackson.object.mapper=org.opendc.web.server.util.QuarkusObjectMapperSupplier
